#CAP-15 update traits for elgamal capsule rug
